poll vote server code finished
