+++ app.js
@@ -230,7 +230,7 @@ app.delete('/json/group/forum/thread/:id', auth(), forums.thread.delete);
 app.post('/json/group/forum/thread/solved', auth(), forums.thread.updateSolved);
 
 // @desc: Update poll
-app.patch('/json/group/forum/thread/poll/:id', auth(), forums.thread.updatePoll);
+app.patch('/json/group/forum/thread/poll/:id', auth(), forums.poll.vote);
 
 // @desc: Create new post in forum thread
 app.post('/json/group/forum/post/create', auth(), forums.post.create);
