+++ server/routes/forums/index.js
@@ -4,3 +4,4 @@ exports.comment = require("./comment");
 exports.post = require("./post");
 exports.thread = require("./thread");
 exports.forum = require('./forum');
+exports.poll = require('./poll');
