+++ server/routes/forums/thread.js
@@ -75,6 +75,7 @@ exports.create = function(req, res) {
 					const poll = {
 						'postId': mainPostId,
 						'threadId': threadId,
+						'groupId': group._id,
 						'options': options,
 						'userIdsVoted': [],
 						'allowMultipleOptions': body.poll.allowMultipleOptions,
@@ -315,65 +316,3 @@ exports.updateSolved = function(req, res) {
 	}).catch(utils.isOwnError, utils.handleOwnError(res));
 };
 
-exports.updatePoll = function(req, res) {
-	const userId = ObjectId(req.user._id);
-	const pollId = ObjectId(req.params.id);
-	const forumId = ObjectId(req.body.forumId);
-	const votes = req.body.votes;
-	
-	// Check if at least one option was given
-	const numChosenOptions = votes.reduce((tot, num) => tot + num);
-	if (numChosenOptions == 0) {
-		res.status(400).send('NO_OPTION_CHOSEN');
-		return;
-	}
-	
-	// Get groupId
-	db.collection('forums').findOneAsync({ '_id': forumId }, { 'groupId': true }).then((groupId) => {
-		return db.collection('group_relations').find({ 'groupId': groupId }, { 'userId': true }).toArrayAsync();
-	}).then((groupMembers) => {
-		// Look for current user in group members array
-		const userIsMember = groupMembers.find(member => member.userId == userId);
-		// Return permission status as boolean
-		return !_.isUndefined(userIsMember);
-	}).then((hasUserPermission) => {
-		// If user has no permission, reject promise
-		if (!hasUserPermission)
-			return Promise.rejectPromiseWithMessage(401, 'NO_PERMISSION');
-		
-		// Get poll
-		return db.collection('forum_polls').findOneAsync({ '_id': pollId });
-	}).then((poll) => {
-		// Look if userId is stored in already voted user ids
-		const userCanVote = poll.userIdsVoted.find(votedUserId => votedUserId == userId);
-		// If userCanVote is undefined, the user is allowed to vote
-		return _.isUndefined(userCanVote);
-	}).then((userCanVote) => {
-		if (!userCanVote)
-			return Promise.rejectPromiseWithMessage(401, 'ALREADY_VOTED');
-		
-		
-	});
-	
-	// X Check if at least one option was given
-	// X Check if user has permission to vote
-	// X Check if user has already voted
-	// Check if multiple options are given and if multiple options are allowed
-	
-	// Store poll options in forum post
-	db.collection('forum_polls').findOneAsync({ '_id': pollId }).then((poll) => {
-		
-		// Increment options counts
-		votes.forEach((vote, index) => {
-			let option = _.findWhere(poll.options, { 'index': index });
-			option.count += vote;
-			console.log(option);
-		});
-		
-		// Add user to voted users
-		poll.userIdsVoted.push(userId);
-		
-		console.log(poll);
-	}).then(res.json.bind(res));
-	
-};
